requirements.txt
confluent-kafka
clickhouse-driver
python-dotenv
cassandra-driver

